requirements.txt
python-deriv-api
websockets
python-dotenv
colorama
supabase
